refactor: one route file for project
